rename dict to compression
