+++ crates/starknet-os/src/hints/compression.rs
